feat(module:tree-select): add the public methods
