+++ components/core/tree/public-api.ts
@@ -2,3 +2,4 @@ export * from './nz-tree-base-node';
 export * from './nz-tree-base.definitions';
 export * from './nz-tree-base.service';
 export * from './nz-tree-service.resolver';
+export * from './nz-tree-base';
